Housing 01: Let's go shopping! (#4)
The first part of a 4 part series on buying a house in Portland Oregon!

I've had a lot of fun writing this series almost a year after buying my house.
It's been nice to reflect on the stress of buying the house, the anguish of spending so much money, and the gratitude I have for the support network in my life that allowed me to make such a big and important life purchase.

Not to say buying a house is perfect, but we'll get to that.
I hope the tone of the pieces articulates that buying a house is _interesting_.
Not _good_ nor _bad_ just _interesting_ and worth reading/writing about!